+++ packages/expo/src/winter/fetch/fetch.ts
@@ -19,6 +19,18 @@ const isRequest = (input: any): input is FetchRequestLike => {
   }
 };
 
+const dangerouslyGetBodyFromRequest = (
+  input: FetchRequestLike | FetchRequestInit | undefined
+): BodyInit | null => {
+  if (input != null && input instanceof Request && '_bodyInit' in input) {
+    // NOTE(@kitten): whatwg-fetch has a hidden property for the body input
+    // TODO(@kitten): We should have our own Request class implementation
+    return (input as any)._noBody !== true ? (input as any)._bodyInit : null;
+  } else {
+    return input?.body ?? null;
+  }
+};
+
 // TODO(@kitten): Do we really want to use our own types for web standards?
 export async function fetch(
   input: string | URL | FetchRequestLike,
@@ -26,7 +38,9 @@ export async function fetch(
 ): Promise<FetchResponse> {
   const initFromRequest = isRequest(input);
   const url = initFromRequest ? input.url : input;
-  const body = init?.body ?? (initFromRequest ? input.body : null);
+  const body =
+    dangerouslyGetBodyFromRequest(init) ??
+    (initFromRequest ? dangerouslyGetBodyFromRequest(input) : null);
   const signal = init?.signal ?? (initFromRequest ? input.signal : undefined);
   const redirect = init?.redirect ?? (initFromRequest ? input.redirect : undefined);
   const method = init?.method ?? (initFromRequest ? input.method : undefined);